test(ui): Fix license conclusion tests
Add one test ID for the license conclusion items in the combo box.
Also change the test logic to use comment field instead of the license,
and apply chaining filter to the locator, to find exactly the correct
item, in case some other tests still haven't deleted their items by the
time this test is run.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>