refactor(api): Add options for purl-cleanup endpoint
By default, the run is now triggered as a dry run, to be able to investigate the API logs what
database update queries would be run. To enable the update queries to the database, the dryRun
option needs to be set to false.

The options also include an option to run all available phases, or it can be configured to perform
only a set of wanted phases at a time.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>