feat(ui): Improve going back to Clearance Library from Main UI
The three subtables of the Clearance Library page are implemented with
the Tab component. As each subtable includes the package names as
clickable links, and clicking the link takes the user to the Main UI,
we want to preserve the tab state if the user returns to the Clearance
Library page with the browser's back button. This is done by adding the
tab as a query parameter to the Clearance Library page URL.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>