fix(ui): Fix showing the existing license conclusions for a file
The dropdown list component showing the existing license conclusions
for a file in the license conclusion form of the Main UI will eventually
get removed/replaced.  While waiting for the refactoring, this commit
fixes the component from overflowing with very long context PURL, which
made it almost impossible to find the delete button for the conclusion.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>