refactor(api): Add hasBulkConclusionId query parameter
When making a request from Clearance UI with bulkConclusionId=null, the null value
is dropped from the query somewhere in the process, and the bulkConclusionId is handled
as undefined, so the database query returns all license conclusions, instead of only the
ones that are not part of a bulk conclusion. In order to work around this, this commit
adds a new boolean query parameter called hasBulkConclusionId.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>