refactor: Use contextPurl as the query param name for clarity
To better distinguish, that the GET license-conclusions and GET license-conclusions/count
endpoints offer results with filtering based on context purl, use contexPurl as the query
parameter name. Filtering based on contextPurl returns only the license conclusions that
were made in that specified context. Filtering based on purl (that will return all license
conclusions affecting the files in that specified package) will be added in the following
commits.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>