refactor(api): Edit GET license-conclusions endpoint
The endpoint can now be used to fetch all license conclusions affecting any files in
specified package regardless of whether those conclusions were made in the context of
that package or not. This commit also adds optionality to filter by contextPurl or
username strictly (exact match) or not (substring match) with the contextPurlStrict
and usernameStrict parameters. This commit resolves do-545.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>