refactor(api): Edit license conclusions count endpoint
The endpoint can now be used to get the count of all license conclusions affecting
any files in specified package regardless of whether those conclusions were made in
the context of that package or not. This commit also adds contextPurlStrict and
usernameStrict parameters for allowing filtering either strictly (exact match) or
not (substring match) on those fields.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>