test(ui): Edit package library test
Edit the package library test, so that it works even when the local testing database
has multiple dos-monorepo packages. Also in case of multiple newer packages, the
dos-monorepo package might not be seen in the first view of the package library,
so this commit also adds searching dos-monorepo with the search input.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>