fix(ui): Fix a lint error about early return
Lint thinks using early return before state handling is violating the
rules of hooks:

'Error: React Hook "useMainUiStore" is called conditionally. React Hooks
must be called in the exact same order in every component render. Did
you accidentally call a React Hook after an early return?'

Fix this by reverting to set the purl and path states to Zustand in
a useEffect.  This commit is written as such, rather than rewriting
history, to be able to easily get back to the issue, if changes are
needed in the future.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>