test(ui): Improve login test
Earlier there was one regression, which prevented login into the
Clearance UI. This was due to the navigation bar not being drawn
properly, and the login test didn't catch it, since it began with
navigating straight to the /login endpoint.  The login test now
begins with navigating to the root endpoint, and then clicking the
login button.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>