test(ui): Fix bulk conclusion tests picking up wrong UI element
With the recent change in the UI to reduce background blurring when a
dialog is open, it was suspected that some e2e tests might start failing
due to picking up UI elements from the background of the dialog, instead
from the dialog itself.  This now happened, as the tests were picking
up the license filter selection combobox from the background UI instead
of the one in the dialog, because they had identical placeholder texts.

The commit fixes the bulk conclusion tests to pick up the correct UI
element by changing the license filter selection placeholder text in
the Main UI, which was needed anyway (the placeholder text was not very
informative).

As a heads up, we can probably improve the e2e tests with more
intelligent locators that constrain the search to the dialog itself,
not to its background.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>