refactor(ui): Rename query state variable
Replace 'purl' query state variable name with 'searchPurl' to prevent an
unhandled runtime error. The error was caused, as other components in the ui
use a query parameter named purl for other things, and a search string is not
necessarily a valid purl, so there was a problem with that search purl being
validated somewhere where it shouldn't have, and throwing an error.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>