test: Edit seeding data and fix tests accordingly
Edit test data, so that the license expressions of license finding matches are SPDX
compliant. While at it, update the seeding data to a newer version of dos, and fix
an issue with the matches being saved with the wrong foreign key license finding id.
Also fix clearance ui tests according to the changes.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>