feat(ui): Change tree selection logic
In order to try and fix the issue with the tree selection in filtering
mode, change the the selection logic a bit.  This didn't solve the
issue, so a separate issue has been created for it.

Also clean up log messages and hide the selection buttons from the UI
for now.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>