feat(ui): Apply whole subpath substitution to the glob pattern
For all the folders selected, substitute the whole subpath with the
glob pattern path/**.  Also reduce [a/b/**, a/b/c/**, a/**] to [a/**],
which is the most general pattern that matches all the selected paths.

Additionally, introduce a function, which will handle the grouping of
the paths that remain after all substitutions, and the creation of the
final glob pattern.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>