feat(ui): Change the way package filetree is filtered
Earlier the filtering of filetree by license was done by calculating a
new filetree from the original, where there were only the nodes left
that matched the selected license finding.

While this works, it is very unefficient and possibly slow, because it
doesn't make use of the search capabilities of react-arborist component.
Moreover, it introduced a bug with the tree selection mode in filtering
state, where the selection status of the nodes was not synced from the
filtering state to the main view.

This commit changes filtering to use the search capabilities of
react-arborist. Filtering is done by searching the filetree either with
freeform text match, or with the selected license finding match, and the
search results are then used to render the filetree.

This way we get rid of calculating and upkeeping two separate filetrees,
so we really only work with one filetree data. For a huge filetree, we
probably also save browser memory and improve the UX.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>