test(ui): Fix bulk conclusion tests
Adding a bulk conclusion leaves the tree in an opened state, and
Playwright tests assumed that the test starts with a closed tree.
This lead to some tests failing, when trying to open some elements for
checking the license conclusion. Fix this by clicking Expand and
then Collapse, so all tree nodes are closed before the test starts.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>