feat(ui): Add a license colour circle to individual match buttons
Add a coloured circle indicating the found license match to all match
buttons.  The colour logic follows the rest of the Main UI: when the
match is a compound license, the circle will be gray, otherwise, it
is the colour used for the individual licenses throughout the UI.

Also tune the button colours to stand up more clearly in both light
and dark themes.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>