refactor(api): Replace keycloak admin client
As it isn't possible to change the api app into an ECMAScript module at least
for now, using the @keycloak/keycloak-admin-client is not possible, see issues
https://github.com/keycloak/keycloak/issues/19829 and https://github.com/keycloak/keycloak-nodejs-admin-client/issues/523,
and the substitute packages don't import the needed classes to be able to properly
handle the possibly thrown errors, therefore will be creating a Keycloak client
with Zodios. These issues have already hindered the development process a lot,
and chances are that the problems won't end there, so possibly it is best to
create an own solution for this as the number of different kinds of needed requests
to the admin-cli is not very big.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>