fix(ui): Don't show more than 10 license hit circles in Main ui
Especially 3rd party notices files can have huge number of license hits
from ScanCode, and showing a hit circle for all of the licese hits
breaks the layout of the filetree. This commit limits the number of
license hit circles to 10, and shows an ellipsis if there are more hits.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>