feat(api): Add temp scanner router
This commit resolves do-595. When switching to Keycloak, the scanner router
only needs changes in handling the authentication of the bearer token in
requests made to some of these endpoints. This token is now an attribute
of the Keycloak user, instead of a field in the User table in the db.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>