feat(api): Add initial temp db_queries file
This is needed, because some of the queries need to be changed so that they
don't use/return the database User table related fields anymore, but rather
the keycloak user id, and the current queries still need to exist as long as
the previous routes are still in use. The copy of the existing db_queries
file is added as the initial file, so that the changes can be better seen in
following commits. This will be a temporary file, the previous db_queries
file can be replaced with this in the process of switching to use Keycloak
completely.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>