feat(api): Add get requests used in Clearance Library LC table
The changes needed in these requests are in the handling of a request that uses
a username filter in the query, and in fetching needed usernames for the result
set. The license conclusion, bulk conclusion, and path exclusion records in the
database will in the near future only have access to the user id, so when
filtering based on username, the corresponding id's have to be fetched first
from the Keycloak server, and used in the database queries finding records.
Also the database results won't include the username anymore, so the usernames
for the results need to be added separately.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>