feat(api): Add LC post, put and delete requests
For the time being, when adding a license conclusion, there will still be two
separate user id's added there, until it is possible to remove the User table
completely from the database. In the update and delete queries, the check that
the user is either on admin or the creator of said license conclusion is now
done using the Keycloak user id and roles.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>