refactor(ui): Simplify convertJsonToTree function
With the changes in backend, converting the JSON object to a
react-arborist tree structure can be simplified, as the excluded files
do not need to be calculated from the glob pattern amymore, but an
isExcluded property is returned instead from the backend for each file
in the tree.

Using this new information provided by the useGetFileTree hook, running
a separate query to get path exclusions can be avoided, which makes the
component state sync better with the UI when in selection mode.

Signed-off-by: Etsija <jyrki.keisala@gmail.com>