refactor(ui): Edit refreshAccessToken handling and calling
Use the zodios client for the request, this makes it more readable and concise. Also change the
condition under which the token will get refreshed to prevent it from being invalid, as the calls
to the API will always need valid access tokens. Also don't expose the refresh token in the returned
session to the client side, as it won't be needed there, and could be a security risk.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>