refactor(api): Remove LicenseRef-dos-exception replacement
The replacement will no longer be needed, as ORT will be able to handle the "AND exception"
type of expression, and furthermore, the fixup is currently causing more work with unnecessary
unhandled license rule violations.

Some background for the initial way of implementing the change in the first place, and why
the computed field will be left here:
The post processing was done with a computed field, so that there wouldn't be a need for
for loops everywhere where license findings are handled, and a computed field could be easily
added to any query that fetches license findings. The original database field was renamed so
that the field could still have the same name as before, and there wouldn't be a need for
changing a bunch of expected types and zod schemas. The computed field is left there for now,
in case this change needs to be reverted, or some other processing is needed for the license
findings in the near future.

Signed-off-by: Johanna Lamppu <johanna.e.lamppu@gmail.com>