refactor(ui): Don't render affected files tooltips when editing
There was a ~3s lag for inputs to become responsive and the cursor being shown
after clicking an input. Returning null for the affected files tooltip when editing
fixes this issue. There is still sometimes a lag when pressing the edit button, and
when pressing the cancel button, until the editable inputs are shown/hidden. So
possibly further investigation into the affected files tooltip is needed to see
if the lag can somehow be avoided.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>