refactor(ui): Edit LC table data handling and onClick events
Looking back at the original example used when implementing the editable rows in
the license conclusions table, noticed that the revertData function with a false
boolean value should have made changes to the initial data of the table component.
For better readability, separating this from the revertData function to its own
updateOriginalData function and adding a new state variable originalData for it.
This might be a bit obsolete, as the initial data will change after saving changes
when the data is fetched again.

Also separating the handling of the edit, save and cancel onClick functions for
better readability.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>