refactor(ui): Remove conditional rendering of affected file tooltips
Interestingly, there is no lagging anymore, so it must have been a glitch on my
local machine when I was implementing the editing for the bulk conclusions. Also
checked the remote, there is no lagging present (there was some for me locally
also when pressing the edit, and the cancel -buttons).

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>