chore(seed): Fix creating matches for test data
Noticed while testing, that some of the matches weren't actually saved
in the local dev database when seeding the test data, and realized the seed
script had a logic error. This resulted in license findings being reported
to ORT with score=null, which then caused an error and stopped the local
ORT run.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>