test(sa): Use superagent in test instead of chai-http
There was an issue with the formidable package "hexoid is not a function"
due to which the formidable version needed to be forced to a prior version.
However, older versions of formidable are now critical vulnerabilities, and
the superagent package now has an update where a new version of formidable
is used that doesn't throw the error anymore. As the functionality in this
test doesn't require anything additional from the chai-http, or the supertest
packages (the dependency chain was chai-http - supertest - superagent -
formidable), changed the test to use the superagent package instead of waiting
for this newer version of the superagent to be taken into use in those other
packages.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>