feat(api): Add onStartUp function
This function will set the state for ScannerJobs that were in processing state
before the server was restarted to 'failed' as the processing was interrupted.
The SystemIssue table is utilized to find out if the same issue occurs for the
same package multiple times, as this implies, that the processing of the package
is what makes the server unresponsive and leads to the server being restarted.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>