fix(schema): Allow single-file BC on back-end
For bulk conclusions, the pattern should be a glob OR a single-file BC.
This was checked correctly on the front-end, but not on the back-end,
where the pattern was always required to be a glob.  This lead to the
front-end form accepting a single-file BC, but the back-end rejecting
it, which lead to a confusing Zodios error being shown to the user.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>