refactor: Edit realm roles claim name in access token
Previously, the realm roles were added to the token with a custom claim name,
but this was not compatible when querying an authorization decision from the
protocol/openid-connect/token endpoint, as it was unable to find the realm roles
associated with the user from the token due to the custom name. So in order to
be able to use the authorization setup for the api client, the claim name is
reset to its original form.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>