refactor(api): Add middleware to authorize requests by permissions
The permissions are configured with Keycloak authorization services. Use the
middleware from keycloak-authorization-services to get a decision from Keycloak
if the user is permitted to access the endpoints. This will allow for a more
flexible setup, where the permissions are not hard-coded for certain roles,
and the permissions for a user/group or a new role can be added/edited "on-the-fly".

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>