refactor(validation-helpers): Add permission request for keycloakAPI
The request is made to the same /protocol/openid-connect/token endpoint as getting or
refreshing an access token, so the request and response schemas are edited so that
they will have union types. This also leads to having to use type casting for the
response from this endpoint to be able to use the response object as needed.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>