refactor(ui): Fetch permissions on useUser hook
These permissions will be used in the components to determine which actions a
user can perform. The request to the Keycloak server is done in a Next API
route so that the Keycloak env variables don't have to be exposed to the browser
and don't need to be replicated with a NEXT_PUBLIC-prefix.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>