refactor(ui): Parse message from error
Only default hard coded error messages, or error.message, which can be quite undescriptive,
were used in some components, so a helper function is added that can be used to parse
the error, as most errors coming from the API should include a custom message. Also the
styling of the error messages is unified in this commit.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>