refactor(ui): Add tooltips for edit and delete buttons
Also add tooltip messages if the button is disabled due to the action not being
permitted for the logged in user. The tooltips are implemented so that you can
give custom messages to use in the tooltip for the cases when the button is disabled
or enabled. The tooltip will be hidden if no message has been given.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>