feat(api): Add function for searching for filetrees matching a pattern
For the types of glob patterns that have so far been used for BC's and PE's,
this function will now utilize PostgreSQL's tilde operator to search with a
regex. With the current setup, the regex created with the glob-to-regexp
package isn't operable for a lot of special characters that could be used in
glob patterns, but these kinds of patterns haven't so far been used, so the
conversion is sufficient as it is for now. If the glob pattern includes some
of these special characters, for which the db query won't find any rows, the
previously used method of looping through all filetree records for the package
will be used instead.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>