refactor: Move pattern schemas to validation-helpers
Move the schemas for bulk conclusion and path exclusion glob patterns to the
validation-helpers package. This way the same schema is used for both form
validation in the UI and the request body parameter validation for the API,
and any changes to the schemas will be reflected on both. Also move the
validateCurlyBracesInGlob helper function and its unit tests to the package.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>