refactor(ui): Use web worker to determine which paths have LFs
Use a web worker in the background to figure out which paths in the file
tree have license findings. The data from the web worker is used to color
the nodes accordingly. This change will allow for the license findings to
be taken away from the file tree query, which will make the initial loading
and drawing of the file tree faster. Also remove some unused helper files.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>