refactor(ui): Edit staletimes for queries
Use staleTime set to Infinity for the filetree query as it no longer returns any
changing data. Also add a long staletime for the license findings query. In practice
this won't change either but potentially at some point there will be rescanning
of files, so there could be new findings.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>