refactor(api): Edit POST scan-results request body
Add declared license to the request packages data. This is needed, as this endpoint also
adds new Package rows to the database. Also with this change, the declared license for
existing packages can be added here. Use a union type in the request to keep the API
compatible with the old version.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>