feat(ui): Change URLs in package details to clickable links
Repository provenances include a `vcs_url` field that contains the URL
to the repository. Same goes for artifact provenances, which include a
`download_url` field.

Whenever a provenance detail includes a `_url` as part of its key, make
the value a clickable link, where the link is opened in a new browser
tab.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>