refactor(api): Edit getting job state in cron job
Previously the jobs in non final state were fetched and the jobs
were processed in the job state query cron job as they were in
the point when the database query was made. However, there was
an incident where the cron job took several seconds and the state
of the job had changed in the meantime and the job ended up in the
saving results phase unnecessarily for a second time. To prevent
this, fetch the state for the job when it is time to start processing
the job in the cron job for loop.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>