refactor(api): Replace transaction for some state changes
The "waiting" and "active" state changes often happen almost
simultaneously, and occasionally the changes result in a deadlock
when a source with multiple packages is being scanned, as state
changes for all the related scanner jobs don't finish within a
transaction before the next state change is attempted to be saved
to the database. The job state query cron job will eventually
clean these states, but as an attempt to prevent this in the
first place, separate the state update for the parent job from
the updating of the child jobs.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>