refactor(api): Edit saving results for file
There are rare occasions where the saving results phase is restarted,
and results for some files have already been saved. Add check for
file's scanStatus before saving results to prevent duplicate results.
This will need to be rethought when allowing multiple results for the
same file.

Also add exceptions on two database queries to do a retry even if
the query has a unique contraint violation error the first time around,
as the error comes from parrallel tasks creating the same file at the
same time, and this shouldn't throw an error and cause the saving
results phase to fail.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>