refactor(s3-helpers): Check if file already exists before uploading
Add a check if the file already exists in the bucket before uploading it.
This was initially omitted as the assumption was that if the file is not
in the database, it won't be in S3 either, but there may be cases where
a run is attempted multiple times, and reuploading the files is
unnecessary.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>