fix: Fix breaking changes from Keycloak version upgrade
Send the username, email, first name and last name in the update user
request when updating a user's attributes as otherwise they would be
nullified, and a missing username results in an error in the request.

Also change the selector of the "Sign in" button in the UI Playwright
tests, as the button is now a button instead of an input.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>