chore: Edit formatting workflow
Use script instead of npx command for formatting check to
remove the need to change the version in the workflow file
when updating prettier. This will ensure that the version
of prettier set up for the project will be used for the
check, whereas removing the version from the npx command
would result in the latest version of prettier being used
for the check.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>