ci(renovate): Pin packageurl-js to v1.2.1
The `PackageURL` function from newer versions of the packageurl-js
package parses the name of an npm package from the purl case-sensitively
only in specific cases for some legacy npm packages. With the current way
that the package info is saved in the database, the generated purl won't
match the purl that came in the request and will result in an error.
Pin the version until this has been otherwise resolved.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>