refactor: Use service account for Keycloak and API connection
Use the client credentials flow [1] for the connection between the API
and Keycloak by obtaining the access token with the client credentials
grant type [2].

Along with this change, the service account roles `manage-users` and
`view-users` have been added for the API client service account in
Keycloak. The other configuration needed, such as enabling client
authentication, was already configured for the client.

This change will also help streamline an upcoming Terraform
configuration for Keycloak.

[1]: https://auth0.com/docs/get-started/authentication-and-authorization-flow/client-credentials-flow
[2]: https://www.oauth.com/oauth2-servers/access-tokens/client-credentials/

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>