refactor(ui): Centralize setting authorization headers for requests
Centralize setting authorization headers for requests by adding an
interceptor for Zodios that adds the headers before the request is made.
As `useSession` from NextAuth.js is React hook-based, but the Zodios
interceptors are not, use a custom shared utility in `authTokenManager.ts`
that holds the token and keeps it updated. A custom provider
`AccessTokenSync` is added to the app and is used to update the token when
the session updates.

This change will make it easier to change the session provider (which is
currently NextAuth.js).

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>