chore: Dockerize Playwright tests and update Playwright workflow
To help resolve issues with an upcoming change that adds a Keycloak Docker
Compose setup for local development, and which will be used in the GitHub
Playwright workflow as well, edit the setup so that the tests in the GitHub
workflow will be run in Docker Compose. This way all the services will be
on the same network, and there won't be a need for tricky workarounds or
larger refactorings to allow the services to communicate with each other.

With this change, also an earlier workaround in the API tests for fetching
the presigned URL needed to be removed.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>