chore: Use the Docker Compose Keycloak instance in Playwright workflow
Use the Docker Compose Keycloak instance for API and Clearance UI in the
Docker Compose environment and use it in the Playwright tests in the
GitHub workflow.

With this change, also the end-to-end user credentials don't need to be
given as environment variables anymore, and the `E2E_USER_USERNAME` and
`E2E_USER_PASSWORD` variables are completely removed.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>