test(api): Use Playwright's `APIRequestContext` for API testing
Refactor the tests to use Playwright's `APIRequestContext` [1] instead of
the Zodios client. The `APIRequestContext` provides ways to also test for
example the response status code, and as this will be needed in some
upcoming tests that test permissions for an endpoint for different user
types, use it throughout the tests for clarity.

The downside of this is that the response body isn't automatically typed,
so for instance the type for the `jobDetails` in the first test needs to
be explicitly declared, as the result is iterated to create a list that
is used for validating the result, and this causes a TypeScript error if
not explicitly declared.

[1]: https://playwright.dev/docs/api/class-apirequest#api-request-new-context

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>