chore(docker): Rename services
Rename the `postgres` service to `db` and the `keycloak` service to `auth`
in order to distinguish these from the services started locally for ORT
Server. The intent is to have the DOS and ORT Server services on the same
network in the local development environment in order to allow these to
communicate with each other, and to enable some testing for combining the
DOS and ORT Server realms in Keycloak.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>