chore: Edit token mapper setup
As the DOS and ORT Server realms are combined in Keycloak, the token
mappers need to be slightly adjusted, as the client roles for the ORT
Server client should not be added to the token, as with a high number of
entities on the server, the amount of these roles is huge, and therefore
will make the token too big. So instead of adding all client roles to the
access and id tokens, only add roles for the clients that are needed.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>