chore: Edit token mapper setup
When having both the DOS and ORT Server services in the same realm in
Keycloak, the token mappers need to be slightly adjusted. The client roles
mapper which contains all roles for all clients won't be added to the access
token or the id token as per ORT Server's requirements. Instead, a separate
mapper for the DOS API client roles is added, which doesn't contain roles
for other clients.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>