refactor(ui): Add handling for 401 responses
If the session gets killed (for example due to a logout in a different
service that is connected to the same authentication), the NextAuth session
won't detect the error until it tries to refresh the access token. This
might take several minutes, and in the meantime the API requests already
start failing.

Add handling for this in the TanStack `queryClient`, so that it is caught
earlier, and the user gets properly redirected to the login page. The
original request will be retried before this error handling kicks in, and
there is an additional query being made to the `userinfo` endpoint to help
mitigate the risk that the failure could be caused by an unlikely but
possible temporary network flakiness.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>