refactor(api): Migrate to API tokens not tied to user accounts
The API tokens that are used to authenticate access to getting scan results,
adding jobs and getting package configurations, used to be personal tokens
for all users. This is changed as there's a need for more restrictions to
the access to these endpoints.

As the number of needed tokens is low and it is expected that they rarely
change, these tokens are added through an environment variable for
simplicity and faster access. However, if the need emerges, `CRUD` support
for token handling for server admins could be added later.

The main motivation for this change comes from developing DOS alongside the
ORT Server, and from an upcoming feature on DOS that will allow users to
create their own clearance items. These clearance items will be linked to
the organization the user belongs in, and for example the package
configurations endpoint will support returning clearance items based on an
organization ID, and so the need to restrict access to this endpoint is
vital, as users shouldn't be allowed to access clearance items on other
organizations than their own.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>