ci(github): Add docker build task in Playwright workflow
The services defined with the `depends_on` attribute in the Docker Compose
file used to be built automatically in the GitHub actions. However, this
behavior seems to have changed, so add a task for building all the images
before the services are started, so that the services that a service
depends on are properly started.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>