feat(database): Add Curator table
Add a table for `Curator` and create new rows for each distinct `userId`
found in the `LicenseConclusion`, `BulkConclusion` and `PathExclusion`
tables, mapping the `remoteId` for the `Curator` to `userId`. The username
for the new curator rows will later on be updated from Keycloak.

Also add `curatorId` columns for the `LicenseConclusion`, `BulkConclusion`
and `PathExclusion` tables and backfill them in the migration. The `userId`
column will be removed in a subsequent commit after switching to use the
curator data instead throughout the code.

The `Curator` table will later on connect the user to a clearance group
that the user can make clearances in. In the future, the table can also
hold some additional information about the curator, such as qualifications.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>