fix(api): Fix logic for handling the saving results state
Previously the assumption was that if a job was found by the `jobStateQuery`
cron job in the `savingResults` state and stayed in that state for longer
than 30 minutes, that the server had likely restarted, and so the saving of
the results was restarted in that case. However, in the case where the
actual saving of the results takes a lot longer, the saving of the results
was actually restarted multiple times until the server got overloaded with
simultaneously handling the large results and crashed due to a "JavaScript
heap out of memory" error.

Change this so that in the case of a server restart, the saving of the
results is restarted in the `onStartUp` function, but only if it hasn't
been restarted before. If there has already been a restart, set the status
to failed, as this indicates that the saving of the results may have
crashed the server. In the `jobStateQuery`, set a threshold for how long
a job can be in the `savingResults` state until it is marked as failed.
Leave the job in the queue for inspection when setting the state to failed.
A cron job for cleaning old jobs from the queue will be added in a
subsequent commit.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>