feat(database): Add ClearanceGroup table and relation tables
Add a table for clearance groups and tables for connecting clearance items
and curators to the group. The clearance groups will allow to group
clearance items from curators so that they can also be retrieved based on
the groups, and so that support for allowing specific tokens to access
data from specific groups can be added.

Signed-off-by: Johanna Lamppu <johanna.lamppu@doubleopen.org>